backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pydantic==1.10.11
aiofiles==23.1.0
python-jose[cryptography] 
passlib[bcrypt]